requirements.txt
pydantic~=2.10.6
pydantic[email]
pydantic-settings
fastapi~=0.115.8
python-dotenv~=1.0.1
uvicorn
python-jwt
PyJWT~=2.8.0
python-jose[cryptography]
passlib[bcrypt]~=1.7.4
python-multipart
bcrypt
sqlalchemy[asyncio]~=2.0.38
asyncpg
psycopg[binary]
fastapi
pydantic_core~=2.27.2
alembic~=1.14.1
jinja2
boto3